Project Proposal
Cryptocurrency Money Making Model

Purpose and Use Cases

The goal in this project is to process the transaction data of Ethereum (a cryptocurrency coin) to suggest a trade action (buy, sell, or no action). 

The users of this model are people who invest in cryptocurrencies such as myself. This model would help the investors know at what price to sell or to buy. If the price remains within the suggested buy and sell range then no trade action would be recommended by this model. 

Trading cryptocurrencies by individual investors is mostly decided based on emotions that do not consider factors such as the number of transactions, transaction volumes, market capacity, generated coins, and more. The decisions are mostly based on the price trend and personal set goals because that is the fastest way to make a decision. 

In this project the goal is to back up the trade decision-making by a solution that could constantly process the Ethereum transaction data. Many investors do not have the time to constantly keep an eye on the prices which could result in losing chances to make money. For example an investor might check the price on a Friday at $200 and then two weeks later again at a price of $200 when the price reached a dip of $180 and a peak of $350 in between the two times. Good opportunities to make money were lost in this period. The goal of this model is to notify the investor of these chances and optimize the trade action recommendation using relevant factors.

Data Sources

I will use the data shared by coinmetrics.io. I will also use more detailed live datasets released by Poloniex (https://poloniex.com/exchange#btc_eth) for this project.

Approach

Data Preparation
Data Wrangling
Data Analytics
Modeling recommendations 
Testing the model on historical data
Applying the model on new data and make transactions accordingly 

Deliverables

The deliverables of this project will be a model that processes CSV and JSON data of the historical Ethereum transactions to specify an upper and a lower limit for buy/sell actions. It will consider inputs such as transaction fees and a sensitivity factor from the user to base the recommendations on.